New commit

+++ fargopy/simulation.py
@@ -8,6 +8,7 @@ import fargopy
 ###############################################################
 import subprocess
 import os
+import re
 
 ###############################################################
 # Module constants
@@ -64,20 +65,81 @@ class Simulation(fargopy.Conf):
         
     def run(self,
             mode='async',
-            options='-m'):
+            options='-m',
+            mpioptions = '-np 1',
+            resume=False
+            ):
         
         if not self._is_setup():
             return
+        if self._is_running():
+            return 
         
         if self.fargo3d_binary is None:
             print("You must first compile your simulation with: <simulation>.compile().")
             return
         
+        # Mandatory options
+        options = options + " -t"
+        if 'run_options' not in self.__dict__.keys():
+            self.run_options = options
+        
         self.logfile = f"{fargopy.Conf.FARGO3D_FULLDIR}/{self.setup_outputs}/{self.setup}.log"
-        self.fargo3d_process = fargopy.Conf._run_fargo3d(f"{self.fargo3d_binary} {options}",
+
+        # Select command to run
+        precmd=''
+        if fargopy.Conf.FARGO3D_PARALLEL:
+            precmd = f"mpirun {mpioptions} "
+
+        self.fargo3d_process = fargopy.Conf._run_fargo3d(f"{precmd}./{self.fargo3d_binary} {options}",
                                                          logfile=self.logfile,
                                                          mode=mode,
+                                                         resume=resume,
                                                          options=self.setup_parameters)
+        
+    def resume(self,since=0,mpioptions = '-np 1'):
+        """Resume a simulation from a given snapshot
+        """
+        if not self._is_setup():
+            return
+        if not self._is_resumable():
+            return 
+        if self._is_running():
+            return 
+        
+        print(f"Resuming from snapshot {since}")
+        self.run(mode='async',mpioptions=mpioptions,resume=True,options=self.run_options+f' -S {since}')
+
+    def _is_resumable(self):
+        outputs_directory = f"{fargopy.Conf.FARGO3D_FULLDIR}/{self.setup_outputs}/"
+        if not os.path.isdir(outputs_directory):
+            print("There is no output to resume at '{outputs_directory.replace('//','/')}'.")
+            return False
+        return True
+    
+    def _get_snapshots(self):
+        if not self._is_setup():
+            return
+        if not self._is_resumable():
+            return
+        
+        error,output = fargopy.Util.sysrun(f"grep OUTPUT {self.logfile}",verbose=False)
+        if not error:
+            snapshots = output[:-1]
+        else:
+            snapshots = []
+        return snapshots
+        
+    def _is_running(self):
+        if self.fargo3d_process:
+            poll = self.fargo3d_process.poll()
+            if poll is None:
+                print(f"The process is already running with pid '{self.fargo3d_process.pid}'")
+                return True
+            else:
+                return False
+        else:
+            return False
 
     def stop(self):
         """Stop the present running process
@@ -105,6 +167,7 @@ class Simulation(fargopy.Conf):
 
     def status(self,
                mode='isrunning',
+               verbose=True
                ):
         """Check the status of the running process
 
@@ -113,29 +176,41 @@ class Simulation(fargopy.Conf):
                 Available modes:
                     'isrunning': Just show if the process is running.
                     'logfile': Show the latest lines of the logfile
+                    'outputs': Show (and return) a list of outputs
+                    'snapshots': Show (and return) a list of snapshots
 
         """
         if not self._is_setup():
             return
         
-        if 'isrunning' in mode:
+        # Bar separating output 
+        bar = f"\n{''.join(['#']*80)}\n"
+        # vprint
+        vprint = print if verbose else lambda x:x
+
+        if 'isrunning' in mode or mode=='all':
+            vprint(bar+"Running status of the process:")
             if self.fargo3d_process:
                 poll = self.fargo3d_process.poll()
 
                 if poll is None:
-                    print("The process is running.")
+                    vprint("\tThe process is running.")
                 else:
-                    print(f"The process has ended with termination code {poll}.")
-
-        if 'logfile' in mode:
-
-            if 'logfile' in self.__dict__.keys() and os.path.isfile(self.logfile):
-                print("The latest 10 lines of the logfile:\n")
-                os.system(f"tail -n 10 {self.logfile}")
+                    vprint(f"\tThe process has ended with termination code {poll}.")
             else:
-                print("No log file created yet")
+                vprint(f"\tThe process is stopped.")
 
-        if 'outputs' in mode:
+        if 'logfile' in mode or mode=='all':
+            vprint(bar+"Logfile content:")
+            if 'logfile' in self.__dict__.keys() and os.path.isfile(self.logfile):
+                vprint("The latest 10 lines of the logfile:\n")
+                if verbose:
+                    os.system(f"tail -n 10 {self.logfile}")
+            else:
+                vprint("No log file created yet")
+
+        if 'outputs' in mode or mode=='all':
+            vprint(bar+"Output content:")
             error,output = fargopy.Util.sysrun(f"ls {fargopy.Conf.FARGO3D_FULLDIR}/{self.setup_outputs}/*.dat",verbose=False)
             if not error:
                 files = [file.split('/')[-1] for file in output[:-1]]
@@ -145,11 +220,26 @@ class Simulation(fargopy.Conf):
                     if ((i+1)%10) == 0:
                         file_list += "\n"
                 file_list = file_list.strip("\n,")
-                print(f"\n{len(files)} available datafiles:\n")
+                vprint(f"\n{len(files)} available datafiles:\n")
                 self.output_datafiles = files
-                print(file_list)
+                vprint(file_list)
             else:
-                print("No datafiles yet available")
+                vprint("No datafiles yet available")
+
+        if 'snapshots' in mode or mode=='all':
+            vprint(bar+"Snapshots:")
+            self.output_snapshots = self._get_snapshots()
+            nsnapshots = len(self.output_snapshots)
+            if nsnapshots:
+                vprint(f"\tNumber of available snapshots: {nsnapshots}")
+                if nsnapshots > 1:
+                    find = re.findall(r'OUTPUTS\s+(\d+)',self.output_snapshots[-2])
+                    self.resumable_snapshot = int(find[0])
+                    vprint(f"\tLatest resumable snapshot: {self.resumable_snapshot}")
+                else:
+                    vprint(f"\tNo resumable snapshot")
+            else:
+                vprint("No snapshots yet")
 
     def clean_output(self):
         """
@@ -157,11 +247,13 @@ class Simulation(fargopy.Conf):
         """
         if not self._is_setup():
             return
+        if self._is_running():
+            return 
         
         print(f"Cleaning simulation outputs...")
         run_cmd = f"rm -r {fargopy.Conf.FARGO3D_FULLDIR}/{self.setup_outputs}/*"
         error,output = fargopy.Util.sysrun(run_cmd,verbose=False)
         if error:
-            print("No output directory has been created yet.")
+            print("Output directory is clean already.")
         if not error:
             print("Done.")
